Move `ty::{OutlivesPredicate, PolyRegionOutlivesPredicate, PolyTypeOutlivesPredicate, RegionOutlivesPredicate, TypeOutlivesPredicate}` to their own little module (cute)
