+++ compiler/rustc_middle/src/ty/predicate.rs
@@ -7,14 +7,19 @@ use rustc_type_ir::WithCachedTypeInfo;
 
 use crate::ty::{
     self, AliasRelationDirection, Binder, BoundConstness, Clause, ClosureKind, CoercePredicate,
-    Const, DebruijnIndex, EarlyBinder, GenericArg, PolyProjectionPredicate,
-    PolyTypeOutlivesPredicate, SubstsRef, SubtypePredicate, Term, Ty, TypeFlags,
+    Const, DebruijnIndex, EarlyBinder, GenericArg, PolyProjectionPredicate, SubstsRef,
+    SubtypePredicate, Term, Ty, TypeFlags,
 };
 
 mod instantiated_predicates;
+mod outlives_predicate;
 mod trait_predicate;
 
 pub use instantiated_predicates::InstantiatedPredicates;
+pub use outlives_predicate::{
+    OutlivesPredicate, PolyRegionOutlivesPredicate, PolyTypeOutlivesPredicate,
+    RegionOutlivesPredicate, TypeOutlivesPredicate,
+};
 pub use trait_predicate::{PolyTraitPredicate, TraitPredicate};
 
 /// Use this rather than `PredicateKind`, whenever possible.
